Created a window manager object.
Methods:

windowManager.new(content,title)
Creates a new window with the content and title supplied to the method.
The created window will equal the size of the content provided to it and have a titlebar with a title and a close button and will be able to be dragged around by the titlebar.

windowManager.close()
Kinda janky, only for use from an onclick="" from the default close button, closes the window which it belonged to.

Variables:
windowManager.count
Is equal to the number of windows opened by the manager, it increases with every window created so it will not always be equal to the number of windows currently open, mostly used to set specific IDs to windows.

Demo:
There is a demo use of this function in a startmenu entry called "New Window", if you view the HTML you can see how it is used, but it can be called from JS just as easily.